refactor: move to use-client/plugin.ts
